style(packages): change toggle style
