+++ packages/babylon-core-ui/src/components/Toggle/Toggle.stories.tsx
@@ -1,5 +1,4 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { FaLock, FaLockOpen } from "react-icons/fa";
 
 import { Toggle } from "./Toggle";
 
@@ -14,7 +13,6 @@ type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
   args: {
-    activeIcon: <FaLockOpen size={10} />,
-    inactiveIcon: <FaLock size={10} />,
+    defaultValue: true,
   },
 };
