+++ packages/babylon-core-ui/src/components/Toggle/Toggle.tsx
@@ -1,5 +1,4 @@
 import { twJoin } from "tailwind-merge";
-
 import { useControlledState } from "@/hooks/useControlledState";
 import "./Toggle.css";
 
@@ -7,22 +6,47 @@ export interface ToggleProps {
   value?: boolean;
   defaultValue?: boolean;
   onChange?: (value: boolean) => void;
-  activeIcon?: JSX.Element;
-  inactiveIcon?: JSX.Element;
+  className?: string;
+  disabled?: boolean;
 }
 
-export function Toggle(props: ToggleProps) {
-  const [value = false, setValue] = useControlledState<boolean>({
-    value: props.value,
-    defaultValue: props.defaultValue,
-    onStateChange: props.onChange,
+export const Toggle = ({
+  value,
+  defaultValue = false,
+  onChange,
+  className = '',
+  disabled = false,
+}: ToggleProps) => {
+  const [isOn = false, setIsOn] = useControlledState<boolean>({
+    value,
+    defaultValue,
+    onStateChange: onChange,
   });
 
+  const handleClick = () => {
+    if (!disabled) setIsOn(!isOn);
+  };
+
   return (
-    <div className="bbn-toggle" onClick={() => void setValue(!value)}>
-      <span className="bbn-toggle-bg">{props.activeIcon}</span>
-      <span className={twJoin("bbn-toggle-control", value && "bbn-toggle-control-active")} />
-      <span className="bbn-toggle-bg">{props.inactiveIcon}</span>
-    </div>
+    <button
+      type="button"
+      onClick={handleClick}
+      disabled={disabled}
+      className={twJoin(
+        'bbn-toggle',
+        isOn && 'bbn-toggle-active',
+        disabled && 'bbn-toggle-disabled',
+        className,
+      )}
+      aria-pressed={isOn}
+      aria-label="Toggle"
+    >
+      <span
+        className={twJoin(
+          'bbn-toggle-control',
+          isOn && 'bbn-toggle-control-active',
+        )}
+      ></span>
+    </button>
   );
-}
+};
